Refactor financial entry forms and APIs for better code reuse
- Created a new FinancialEntryForm component to handle both income and expense entries, reducing code duplication.
- Introduced a base schema and interface for financial entries to unify income and expense models.
- Updated API utilities to use a shared function for fetching and managing financial entries.
- Simplified the Add Income and Expense pages by integrating the new FinancialEntryForm component.
- Enhanced error handling and user preference management for form submissions.
